R4emove caching of widgets in ContaoWidgetManager.
This is needed, as the manipulate widget events might need the updated
data within the model.
We can't move the creation of widgets "after" updating the model, as the
input data has to be fed through the widgets to update the model causing
a chicken egg problem then.
So we now settle for creating the widget instances twice which solves
the problem.