Remove pid handling from Controller
The PID handling is done via EnforceModelRelationshipEvent listeners for
a long time now and therefore it is not needed to apply the conditions
twice. In fact, this produces more problems than it solves, as this
generic solution was not aware of "into" and "after" parameters like the
tree mode listener knows.