Fix Model rendering from callback
The dca callbacks will return a string which potentially contains
percent signs in the label content (like for our encoded ids in URLs).
This confuses the vsprintf() call in FormatModelLabelSubscriber as
something like '%s %3A%3A' will be handled as format string and we
therefore have too few parameters.
We now cirvumvent this by prefixing any invalid format string with an
additional %-sign to omit interpreting these as argument.