Change the handling of tl_limit for the DefaultLimitElement
Behavior:
The value "tl_limit" of the limit field will be handelt as "all". 

Problem: 
If you set up a filter with no result the selected value of the limit field will be 'tl_limit' if you now reset the filter back, the limit will still be "tl_limit" which mean i will get all result after a filter reload. So if you have 20k entries in the list, the system isn't able to display this data. 

Fix: 
I change the handling that the value "tl_limit" will be handeled as nothing. So the system can handle the default handling.